+++ src/entries/toc.ts
@@ -1,0 +1,6 @@
+/**
+ * Subpath export for TableOfContents.
+ * Use for smaller bundles when you only need an on-this-page rail.
+ */
+export { TableOfContents } from "../components/custom/table-of-contents"
+export type { TOCItem, TableOfContentsProps } from "../components/custom/table-of-contents"